@@ -16,20 +16,21 @@
         }
     ],
     "require": {
-        "php": "~7.4.0 || ~8.0.0 || ~8.1.0",
+        "php": "~8.1.0 || ~8.2.0",
         "laminas/laminas-servicemanager": "^3.11",
         "mezzio/mezzio-twigrenderer": "^2.9",
         "laminas/laminas-view": "^2.20",
-        "laminas/laminas-authentication": "^2.10"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^9.5",
-        "squizlabs/php_codesniffer": "^3.6",
+        "laminas/laminas-authentication": "^2.10",
         "dotkernel/dot-navigation": "^3.1",
         "dotkernel/dot-flashmessenger": "^3.1.1",
         "laminas/laminas-form": "^3.1.1",
         "doctrine/doctrine-module": "^5.1",
-        "mezzio/mezzio-authorization": "^1.4"
+        "dotkernel/dot-authorization": "^3.3"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^10.2",
+        "laminas/laminas-coding-standard": "^2.5",
+        "vimeo/psalm": "^5.13"
     },
     "autoload": {
         "psr-4": {
@@ -40,5 +41,22 @@
         "psr-4": {
             "DotTest\\Twig\\": "test/"
         }
+    },
+    "config": {
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
+    },
+    "scripts": {
+        "check": [
+            "@cs-check",
+            "@test"
+        ],
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "test": "phpunit --colors=always",
+        "test-coverage": "phpunit --colors=always --coverage-clover clover.xml",
+        "static-analysis": "psalm --shepherd --stats"
     }
 }
