@@ -1,45 +1,28 @@
 <?php
-/**
- * @see https://github.com/dotkernel/dot-twigrenderer/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-twigrenderer/blob/master/LICENSE.md MIT License
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Twig\Extension;
 
+use Laminas\Authentication\AuthenticationServiceInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
-use Laminas\Authentication\AuthenticationServiceInterface;
 
-/**
- * Class AuthenticationExtension
- * @package Dot\Twig\Extension
- */
 class AuthenticationExtension extends AbstractExtension
 {
-    /** @var AuthenticationServiceInterface */
-    protected $authentication;
+    protected AuthenticationServiceInterface $authentication;
 
-    /**
-     * AuthenticationExtension constructor.
-     * @param AuthenticationServiceInterface $authentication
-     */
     public function __construct(AuthenticationServiceInterface $authentication)
     {
         $this->authentication = $authentication;
     }
 
-    /**
-     * @return string
-     */
     public function getName(): string
     {
         return 'dot-authentication';
     }
 
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new TwigFunction('hasIdentity', [$this, 'hasIdentity']),
@@ -52,7 +35,7 @@ class AuthenticationExtension extends AbstractExtension
         return $this->authentication->hasIdentity();
     }
 
-    public function getIdentity()
+    public function getIdentity(): mixed
     {
         return $this->authentication->getIdentity();
     }
